Day1/5_mlflow_project/requirements.txt
mlflow
pytorch
torchmetrics
